refactor: cast monetary values to integers and update discount calculations and shift reporting logic

+++ app/Models/Payable.php
@@ -23,8 +23,8 @@ class Payable extends Model
     ];
 
     protected $casts = [
-        'total' => 'float',
-        'paid' => 'float',
+        'total' => 'integer',
+        'paid' => 'integer',
         'due_date' => 'date',
     ];
 
@@ -58,9 +58,9 @@ class Payable extends Model
         return $this->warehouse_id ?? $this->purchaseOrder?->warehouse_id;
     }
 
-    public function getRemainingAttribute(): float
+    public function getRemainingAttribute(): int
     {
-        return max(0, ($this->total ?? 0) - ($this->paid ?? 0));
+        return (int) max(0, ($this->total ?? 0) - ($this->paid ?? 0));
     }
 
     public function getAgingBucketAttribute(): string
